Only do relative requires when path starts with .
There are many other ways to require files which I didn't consider in #2866

I tried doing (Pathname.new foo).absolute? but sadly that doesn't detect gems or URLs. I think this simple . prefix is a
good balance between all the other methods and allowing relative includes. I can't think of a way you would do relative
includes without starting with a ..

This should be released as a 0.38.1 soon since 0.38.0 breaks people with non-relative requires.

Fixes #2937
