+++ README.md
@@ -1017,7 +1017,8 @@ require:
 
 Note: The paths are directly passed to `Kernel.require`.  If your
 extension file is not in `$LOAD_PATH`, you need to specify the path as
-relative path prefixed with `./` explicitly, or absolute path.
+relative path prefixed with `./` explicitly, or absolute path. Paths 
+starting with a `.` are resolved relative to `.rubocop.yml`.
 
 ### Custom Cops
